+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.7.17] 2021-10-10
 ### Added
 * API: new order type "changetimestamp"
+* API: streaming servers
+* DB: streaming servers
+* CHECK: streaming servers
 
 ## [0.7.16] 2021-09-29
 ### Fixed